backend/requirements.txt
flask
flask-cors
python-dotenv
pillow
requests
supabase
psycopg2-binary
pytest
protonmail-api-client